requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.2
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.4.0
scikit-learn==1.2.2
scipy==1.10.1
svg.path==6.2
threadpoolctl==3.1.0
Werkzeug==2.2.3
